Refactor default model modal
