+++ src/Settings/SettingsView.ts
@@ -3,8 +3,6 @@ import {
 	App,
 	ButtonComponent,
 	DropdownComponent,
-	Modal,
-	Notice,
 	PluginSettingTab,
 	Setting,
 } from "obsidian";
@@ -13,55 +11,9 @@ import { models, modelNames } from "utils/models";
 import { claudeSonnetJuneModel, openAIModel, geminiModel, GPT4All } from "utils/constants";
 import logo from "assets/LLMguy.svg";
 import { FAB } from "Plugin/FAB/FAB";
+import DefaultModelModal from "Plugin/Components/DefaultModelModal";
+
 const fs = require("fs");
-
-// TODO - abstract to its own component file.
-class DefaultModelModal extends Modal {
-	plugin: LLMPlugin;
-	private defaultModel: string;
-
-	constructor(app: App, defaultModel: string, plugin: LLMPlugin) {
-		super(app);
-		this.defaultModel = defaultModel;
-		this.plugin = plugin
-	}
-
-	handleChangeDefaultModel() {
-		changeDefaultModel(this.defaultModel, this.plugin);
-		const defaultModelSelector = document.querySelector('.default-model-selector > .setting-item-control > .dropdown') as HTMLSelectElement;
-		defaultModelSelector.value = this.defaultModel
-	}
-
-	onOpen() {
-		const { contentEl } = this;
-		contentEl.empty(); // Clear existing content if needed
-
-		// Create the modal content
-		contentEl.createEl('h2', { text: 'Set Model as Default' });
-		contentEl.createEl('p', { text: 'Would you like to set this model as your default model?' });
-
-		// Create the interactive button
-		const yesButton = contentEl.createEl('button', { text: 'Yes' });
-		yesButton.onclick = () => {
-			this.handleChangeDefaultModel()
-			new Notice('Model set as default!');
-			this.close(); // Close the modal after the action
-		};
-
-		// User elects to not save the model as default
-		const noButton = contentEl.createEl('button', { text: 'No' });
-		noButton.style.marginLeft = '16px';
-		noButton.onclick = () => {
-			new Notice('API key saved');
-			this.close(); // Close the modal after the action
-		};
-	}
-
-	onClose() {
-		const { contentEl } = this;
-		contentEl.empty(); // Clear content when the modal is closed
-	}
-}
 
 export default class SettingsView extends PluginSettingTab {
 	plugin: LLMPlugin;
